+++ src/main/kotlin/glsl/plugin/language/GlslFile.kt
@@ -2,8 +2,10 @@ package glsl.plugin.language
 
 import com.intellij.extapi.psi.PsiFileBase
 import com.intellij.openapi.fileTypes.FileType
+import com.intellij.openapi.fileTypes.LanguageFileType
 import com.intellij.openapi.util.IconLoader
 import com.intellij.psi.FileViewProvider
+import javax.swing.Icon
 
 
 /**
@@ -28,3 +30,30 @@ object GlslIcon {
     val PLUGIN_FILE_ICON = IconLoader.getIcon("icons/file-icon.svg", GlslIcon::class.java)
 }
 
+/**
+ *
+ */
+class GlslFileType : LanguageFileType(GlslLanguage.INSTANCE) {
+
+    /**
+    *
+    */
+    override fun getName(): String = "Glsl File"
+
+    /**
+    *
+    */
+    override fun getDescription(): String = "OpenGL shading language file"
+
+    /**
+    *
+    */
+    override fun getDefaultExtension(): String {
+        return "glsl"
+    }
+
+    /**
+    *
+    */
+    override fun getIcon(): Icon = GlslIcon.PLUGIN_FILE_ICON
+}